backend/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
mysql-connector-python==8.1.0
requests==2.31.0